chore: Add admin middleware all dashboard routes *small tweak seeders

+++ routes/web.php
@@ -10,7 +10,7 @@ use App\Http\Controllers\SessionController;
 use App\Http\Controllers\LanguageController;
 use App\Http\Controllers\AdminQuoteController;
 use App\Http\Middleware\CheckLocale;
-
+use GuzzleHttp\Middleware;
 
     // Log In
 
@@ -23,19 +23,19 @@ use App\Http\Middleware\CheckLocale;
 
    // static dashboard
 
-    Route::view('dashboard', 'dashboard')->name('dashboard');
+    Route::view('dashboard', 'dashboard')->name('dashboard')->Middleware('admin');
 
 
 
     // admin movies
 
 
-	Route::get('/movies', [AdminController::class, 'index'])->name('movies');
-	Route::get('/movies/create', [AdminController::class, 'create'])->name('movie.create');
-	Route::post('/movies', [AdminController::class, 'store'])->name('movie.store');
-	Route::delete('/movies/{movie}', [AdminController::class, 'destroy'])->name('movie.destroy');
-	Route::patch('/movies/{movie}', [AdminController::class, 'update'])->name('movie.update');
-	Route::get('/movies/{movie}/edit', [AdminController::class, 'edit'])->name('movie.edit');
+	Route::get('/movies', [AdminController::class, 'index'])->name('movies')->Middleware('admin');;
+	Route::get('/movies/create', [AdminController::class, 'create'])->name('movie.create')->Middleware('admin');;
+	Route::post('/movies', [AdminController::class, 'store'])->name('movie.store')->Middleware('admin');;
+	Route::delete('/movies/{movie}', [AdminController::class, 'destroy'])->name('movie.destroy')->Middleware('admin');;
+	Route::patch('/movies/{movie}', [AdminController::class, 'update'])->name('movie.update')->Middleware('admin');;
+	Route::get('/movies/{movie}/edit', [AdminController::class, 'edit'])->name('movie.edit')->Middleware('admin');;
 
 
 
@@ -43,12 +43,12 @@ use App\Http\Middleware\CheckLocale;
     // admin quotes
 
 
-	Route::get('/quotes', [AdminQuoteController::class, 'index'])->name('quotes');
-	Route::get('/quotes/create', [AdminQuoteController::class, 'create'])->name('quote.create');
-	Route::post('/quotes', [AdminQuoteController::class, 'store'])->name('quote.store');
-	Route::delete('/quotes/{quote}', [AdminQuoteController::class, 'destroy'])->name('quote.destroy');
-	Route::patch('/quotes/{quote}', [AdminQuoteController::class, 'update'])->name('quote.update');
-	Route::get('/quotes/{quote}/edit', [AdminQuoteController::class, 'edit'])->name('quote.edit');
+	Route::get('/quotes', [AdminQuoteController::class, 'index'])->name('quotes')->Middleware('admin');;
+	Route::get('/quotes/create', [AdminQuoteController::class, 'create'])->name('quote.create')->Middleware('admin');;
+	Route::post('/quotes', [AdminQuoteController::class, 'store'])->name('quote.store')->Middleware('admin');;
+	Route::delete('/quotes/{quote}', [AdminQuoteController::class, 'destroy'])->name('quote.destroy')->Middleware('admin');;
+	Route::patch('/quotes/{quote}', [AdminQuoteController::class, 'update'])->name('quote.update')->Middleware('admin');;
+	Route::get('/quotes/{quote}/edit', [AdminQuoteController::class, 'edit'])->name('quote.edit')->Middleware('admin');;
 
     // pages
